ctype: towupper/towlower must respect locale
The C locale only supports ASCII. Fail conversions for values beyond
that range.

This also restructures the wchar case conversion code so that
towctrans_l is not the lowest-level function as that causes both
conversion functions to always be linked in.
